Move buildStakeTree to txhelpers package.
Remove stray continue in rebuilddb.
